+++ src/ui/dialogs/thermogram.h
@@ -89,11 +89,12 @@ private:
     void UpdateExpTable();
     void UpdateDilTable();
     void ExportData();
+    void ImportRow();
 
     PeakPick::spectrum LoadITCFile(QString& filename, std::vector<PeakPick::Peak>* peaks, qreal& offset);
     PeakPick::spectrum LoadXYFile(const QString& filename);
 
-    QPushButton *m_exp_button, *m_dil_button, *m_refit, *m_export_data;
+    QPushButton *m_exp_button, *m_dil_button, *m_refit, *m_export_data, *m_import_row;
     QCheckBox *m_remove_offset, *m_UseParameter, *m_constantVolume;
     QLineEdit *m_exp_file, *m_dil_file, *m_injct, *m_exp_base, *m_dil_base, *m_CellVolume, *m_CellConcentration, *m_SyringeConcentration, *m_Temperature;
     QComboBox* m_scale;